% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_replace_values.R
\name{xml_replace_values}
\alias{xml_replace_values}
\title{Replace values in a xml template document}
\usage{
xml_replace_values(xml, replacement)
}
\arguments{
\item{xml}{a xml document read with \code{read_xml}}

\item{replacement}{one of a one row dataframe, a named list or a named vector where (col)names correspond to  the values to replace in the xml template and the elements are the values to replace with}
}
\value{
a new xml document with changes applied to the xml template
}
\description{
Replace values in a xml template document
}
\examples{
library(xml2)
template <- read_xml("<parent><child nom=\\"child1\\"><grandchild>grandchild1</grandchild>
<grandchild>grandchild2</grandchild></child><child nom=\\"child2\\">null</child></parent>")
toreplace <- c(child1 = "Ben", child2 = "Sarah", grandchild1 = "Brad", grandchild2 = "Jude")
newxml <- xml_replace_values(xml = template, replacement = toreplace)
}
